# doc/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CommandAck": {
        "description": "The reply to every command POST.\n\nDeliberately NOT the resulting state. That arrives on whichever socket the\npage has open, on the server's own schedule, so state has exactly one path\nand there is no ordering for a client to reconcile between two of them. What\nthis carries is only \"the command was understood and applied, and here is\nwhat it was\" -- enough to log and to assert on, and nothing a page renders.",
        "properties": {
          "applied": {
            "title": "Applied",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "applied"
        ],
        "title": "CommandAck",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RangeBody": {
        "description": "A half-open range ``[start, end)`` of the provider's timeline.",
        "properties": {
          "end": {
            "description": "Exclusive end, ISO 8601; must be after start.",
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "start": {
            "description": "Inclusive start, ISO 8601.",
            "format": "date-time",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "RangeBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
...
    "title": "p-SWAMP api",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/time-series-explorer/count": {
      "post": {
        "description": "Count the frames in ``[start, end)``. The batch case: the row-count module\nasks the provider for the range itself, unpaced, and publishes one result\ncarrying this command's request id.",
        "operationId": "time_series_explorer_count",
        "parameters": [
          {
            "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
            "in": "query",
            "name": "client_id",
            "required": true,
            "schema": {
              "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
              "pattern": "^\\d{1,20}$",
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RangeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The range lies outside the provider's coverage, or there is none."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Count",
        "tags": [
          "time-series-explorer"
        ]
      }
    },
    "/api/time-series-explorer/playback/play-range": {
      "post": {
        "description": "Replay exactly ``[start, end)`` at real time, then end paused. The stream\ncase: the player asks the provider for the range and paces it.",
        "operationId": "time_series_explorer_play_range",
        "parameters": [
          {
            "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
            "in": "query",
            "name": "client_id",
            "required": true,
            "schema": {
              "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
              "pattern": "^\\d{1,20}$",
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RangeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The range lies outside the provider's coverage, or there is none."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Play Range",
        "tags": [
          "time-series-explorer"
        ]
      }
    },
    "/api/time-series-explorer/playback/stop": {
      "post": {
        "description": "Pause the replay where it is.",
        "operationId": "time_series_explorer_stop",
        "parameters": [
          {
            "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
            "in": "query",
            "name": "client_id",
            "required": true,
            "schema": {
              "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
              "pattern": "^\\d{1,20}$",
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The range lies outside the provider's coverage, or there is none."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop",
        "tags": [
          "time-series-explorer"
        ]
      }
    },
    "/api/time-series-explorer/refresh": {
      "post": {
        "description": "Ask the provider again what it holds. The one command a page sends when\nthe provider stopped answering: if it is back, the next state carries its\ncoverage and clears the error; if not, the state still says so.",
        "operationId": "time_series_explorer_refresh",
        "parameters": [
          {
            "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
            "in": "query",
            "name": "client_id",
            "required": true,
            "schema": {
              "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
              "pattern": "^\\d{1,20}$",
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "time-series-explorer"
        ]
      }
    },
...
  },
...
}